requirements.txt
pyairbyte
dbt-core